requirements.txt
Django>=1.11.0,<1.12.0
djangorestframework>=3.9.4,<3.10.0
